Expand on test for member already having maximum number of books checked out.
The test now actually sets up the mock repository to return a list of book
loan records that is exactly equal to the maximum number of outstanding book
loans allowed.

I chose to specify all the fields of the book loan, but I could have just
returned a list of empty instances.  I may switch back to returning empty
instances in a future step because returning populated instances may distract
readers of the test from what is really important in the test, which is the
_number_ of instances returned, not the content of the instances.

Another perspective on this may be that returning the instances with somewhat
realistic content (such as WasReturned set to false and a DueDate in the past)
make the test less brittle to sanity checks that may be valid for the code to
do.  I think this perspective is reasonable idea as well.  Like many things
it comes down to weighing trade-offs and making a decision.  It may be worth
considering how much control we have over GetOutstandingBookLoansForMemberAsync
and how likely it is to break.  If we have integration tests elsewhere for
GetOutstandingBookLoansForMemberAsync that run against the real backend, it
is likely reasonable to not put sanity checks in the production code for
it misbehaving badly.