Make the test for disallowing checkouts with too many outstanding loans build.
Once the code is made to build, the test passes.  Commenting out the throw line
will make it fail.  It is somewhat unrelated but I added an await on
Task.CompletedTask so I could leave the method marked as async and not get
warnings.

Now, it may seem cheap that I just am unconditionally throwing here.  However,
making test pass in silly ways like this helps prompt me to think about what
test to write next to fix the obviously wrong implementation.

It may also seem inefficient to do this.  However, how much type did I really
spend writing that one line and running the test?  I normally wouldn't be
committing and writing very detailed commit messages after each of these steps
like I am doing for this example, so the turn around can be very quick.  The
real time will be spent writing the next test, which needs to be written
anyway.

I do think there is a limit to the "level of silliness" that is reasonable to
do to make a test pass.  For example, when testing something like a Fibonacci(n)
function it would be unreasonable to implement it as a switch statment such
that it only works for the inputs the test gives.  I think that a good limit on
the "level of silliness" is thinking about if it is guiding you to write the next test about a real requirement.  I think that once the test suite requires the
function to work for multiple different input values it is unreasonable to do
any silly hard coding to only work for those input values.

Another way to think about it is that unit tests don't need to protect against
code that is so obviously flawed that it would not pass code review.