Test that having fewer than the max outstanding loans does not throw.
I quickly created a copy of the first test that that will make me implement
a real check on the return value of GetOutstandingBookLoansForMember.

There is a ton of code duplication between the two tests right now and it will
be cleaned up, but for now I just decided to make more progress on the
implementation.

There could be some concern around this test not being complete as well.
Afterall in the final implementation we would need to stub out more things in
order to get the CheckoutBookAsync function we are testing to not throw.  That
is true, but not a problem.  It is okay for the individual tests to evolve
and even be deleted as we go.  Potentially deleting tests can sound wasteful,
but often enough is learned by writing them that needed to be learned anyway
that it is not too big of a deal.