Throw only if the member has too many outstanding book loans.
Notice that this is the first time I actually needed to use the
IBookLoanRepository that is passed into the constructor as a dependency.  Until
now I got by with just ignoring that argument entirely.

While writing this code I first tried to do a silly thing by hardcoding the
memberId argument to an empty string.  However, that wouldn't work because the
mock only returns the correct stub value if it gets the correct argument.  I
then changed to to actually pass in memberId.  After this I realized I _could_
cheat and hardcode the argument to "test-member-id".  So I did.

You can make a very valid case that this is code that would not survive code
review.  It would be reasonable to not take this step, but passing the correct
arguments to dependencies is a responsibility of the code and it _can_ be
tested more strongly that in it currently is without much effort.  So at least
for purposes of exploring the limits of TDD and unit testing I decided to do it.