Write test for PastDueBooksException and refactor helpers to share code.
When writing this test I realized that my other tests were actually brittle
because they had book loans expiring in the past and in that situation it would
be valid to throw either the PastDueBooksException or the
TooManyCheckedOutBooksException.  The simplest way to address this would be do
modify the existing helper to return a date in the future.  However, the test
regarding the PastDueBooksException needed to generate outstanding book loans
that are both past due and due in the future.  Because of this it was a good
opportunity to refactor again.  To avoid adding a bunch of arguments and
complexity to SetupOutstandingBookLoansForMember I chose to split the function
into two composable functions.  SetupOutstandingBookLoansForMember now takes a
collection of book loans to return and a new function
CreateOutstandingBookLoans builds a collection of BookLoans with the given
member id and due date.  LINQ makes it easy to build a single collection from
multiple calls to this helper using the Concat method.  I originally called
this method CreateOutstandingBookLoansForMemberWithDueDate, but decided that
the name was a little long and since the call sites would be passing in
well-named constants and variables the long name was not adding a lot of value
so I shortened it.