Fix up success test case set up to have an available copy to it does not throw.
I also considered if this test should just assert that
TooManyCheckedOutBooksException is not thrown.  That keeps the test only
asserting one concern, and is an interesting idea, but it is not something I
generally do.  I think it will be better to evolve this test case into a
general "happy path" test case where all the checks pass.  A test case like
this will be needed anyway.  It may be necessary to have multiple "happy
path" test cases that assert different concerns about the happy path.