Change BookLoan to a C# 9 record with init-only properties and remove Id property
C# 9 records automatically implement value equality which makes them easier to
use from tests without needing to write boilerplate code to either manually
implement equality or lambdas in Arg.Is checks in tests to verify all
properties.  Records also automatically implement ToString which makes test
failure messages (such as those produced by NSubstitute's Received() checks)
easier to read.

The Id property was removed because the business layer doesn't care about it
(at least not yet).  An implementation of IBookLoanRepository may very well
persist records that include an additional Id property, but currently that is
an implementation detail of the persistence layer so there is no need to have
the property in the Business/domain model of a BookLoan.