Don't throw if a subsequent call to TryCreateBookLoan returns true.
Note this code is not correct yet, it keeps calling TryCreateBookLoan with the first
copy id until it succeeds, which in practice would not happen.  We need more tests to
ensure subsequent calls to TryCreateBookLoan have the correct arguments and that if
TryCreatBookLoan returns false for all copies the function does terminate by throwing
NoCopiesAvailableException.