Make test for reminder service not being called pass (in an unreasonable way)
This is further than I would normally go to make a test pass without the code really
being correct, but I also write a test to catch this error.  The new test checks call
order.  I generally avoid checking call order when it does not matter, but in this case
it is actually reasonable to require that the call to TryCreateBookLoan comes before
the call to ScheduleRemindersAsync.